slight fix

+++ ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: lcozdenm <loic.cozdenmat@gmail.com>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/17 01:31:03 by lcozdenm          #+#    #+#             */
-/*   Updated: 2022/09/20 20:32:33 by lcozdenm         ###   ########.fr       */
+/*   Updated: 2022/11/21 15:24:48 by lcozdenm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 size_t	ft_strlen(const char *s)
 {
-	int	size;
+	size_t	size;
 
 	size = 0;
 	while (s[size])
